fix: go through dark styles
Hey! I made these changes because I want to have as small of a blast radius as I can with this PR. The reason I went with tailwind typography is because you can style things nicely without having styles on every element in your markup. I would love it if you could go through each of the remaining files and double check that they actually are required.