some elements changed to dark mode (#2)
* some elements changed to dark mode

* make words use white text

* fix logo for dark mode

* fix background

* fix verses, passageSelector, other pages

* history changes (untested)

* feat: add dark mode for navigation and /history

* feat: update svg to work in dark mode

* fix scrollbar

* remove package-lock.json

* make emaillink border-white in dark mode

* move dark mode within EmailLink

* remove dark:text-white

* remove div background color

* tailwind typography dark

* copyright dark:prose-invert

* add tailwind typography to arena (not sure if working)

* fix: go through dark styles

Hey! I made these changes because I want to have as small of a blast radius as I can with this PR. The reason I went with tailwind typography is because you can style things nicely without having styles on every element in your markup. I would love it if you could go through each of the remaining files and double check that they actually are required.

* remove unnecessary dark:text-white

* fix: go through dark theme again

---------

Co-authored-by: Joshua Wootonn <joshuawootonn@gmail.com>